Move ISO2ROI code into own file
